Work in progress
So far this thing is connecting to discord, using micro and receiving what it needs to perform it's job.

Sample yaml file is included and due to how discord's api goes through a gateway the discord 'guild id' will be needed.  This bot will NOT be able to be shared (which is perfectly fine)